refactor(core/prodtest): Put patch_and_update function to separate source file
