+++ core/.changelog.d/4462.added
@@ -1,0 +1,1 @@
+STWLC38 nvm patch and config update from host
